Scene class implemented according to diagram
Implemented a Scene class

The Scene class has a constructor for taking in a list of world game objects and screen game objects. The destructor should delete the unique pointers and then clear the lists. The Update, DrawWorld, and DrawScreen function implementations are commented out since GameObjects do not yet exist.

Resolves:
http://btechgames.bcit.ca/redmine/issues/7138